💄 style(ui): 调整页面样式和布局 - 修改 body 的 padding 和 margin 属性 - 确保页面元素的视觉一致性 - 更新行间距和列间距的处理方式

+++ host/index.jsx
@@ -70,8 +70,8 @@ function handleNegatePosition(){
 function handleArrange() {
     console.log("Arrange button clicked");
     var columns = parseInt(columnsInput.value) || 3;
-    var rowGap = parseInt(rowGapInput.value) || 10;
-    var colGap = parseInt(colGapInput.value) || 10;
+    var rowGap = parseFloat(rowGapInput.value) || 10;
+    var colGap = parseFloat(colGapInput.value) || 10;
     if (
         useUniformWidthCheckbox.checked &&
         parseFloat(uniformWidthInput.value) < 0
